requirements.txt
asgiref==3.7.2
asttokens==2.4.1
bcrypt==4.0.1
chardet==5.2.0
decorator==5.1.1
Django==4.2.7
django-jazzmin==2.6.0
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
executing==2.0.1
inflection==0.5.1
ipdb==0.13.13
ipython==8.17.2
jedi==0.19.1
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pillow==10.2.0
prompt-toolkit==3.0.41
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.1
pytz==2023.3.post1
PyYAML==6.0.1
reportlab==4.1.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
traitlets==5.13.0
typing_extensions==4.8.0
uritemplate==4.1.1
wcwidth==0.2.10
